move container props to bokeh.core.property.container
