+++ bokeh/core/properties.py
@@ -68,8 +68,6 @@ Basic Properties
 Container Properties
 --------------------
 
-{container_properties}
-
 DataSpec Properties
 -------------------
 
@@ -108,7 +106,6 @@ import logging
 logger = logging.getLogger(__name__)
 
 import base64
-import collections
 import datetime
 import dateutil.parser
 from functools import wraps
@@ -117,18 +114,16 @@ import numbers
 import re
 
 import PIL.Image
-from six import string_types, iteritems
+from six import string_types
 
 from .. import colors
 from ..util.dependencies import import_optional
-from ..util.serialization import convert_datetime_type, convert_timedelta_type, decode_base64_dict, transform_column_source_data
+from ..util.serialization import convert_datetime_type, convert_timedelta_type
 from ..util.string import nice_join, format_docstring
 
-from .property.bases import ContainerProperty, DeserializationError, ParameterizedProperty, Property, PrimitiveProperty
-from .property.descriptors import (ColumnDataPropertyDescriptor, DataSpecPropertyDescriptor,
-                                   UnitsSpecPropertyDescriptor)
+from .property.bases import DeserializationError, ParameterizedProperty, Property, PrimitiveProperty
+from .property.descriptors import DataSpecPropertyDescriptor, UnitsSpecPropertyDescriptor
 from .property.instance import Instance
-from .property.wrappers import PropertyValueColumnData, PropertyValueDict, PropertyValueList
 from . import enums
 
 pd = import_optional('pandas')
@@ -1129,273 +1124,13 @@ class TimeDelta(Property):
 # Container properties
 #------------------------------------------------------------------------------
 
-class Seq(ContainerProperty):
-    ''' Accept non-string ordered sequences of values, e.g. list, tuple, array.
-
-    '''
-
-    def __init__(self, item_type, default=None, help=None):
-        self.item_type = self._validate_type_param(item_type)
-        super(Seq, self).__init__(default=default, help=help)
-
-    def __str__(self):
-        return "%s(%s)" % (self.__class__.__name__, self.item_type)
-
-    @property
-    def type_params(self):
-        return [self.item_type]
-
-    def from_json(self, json, models=None):
-        if json is None:
-            return None
-        elif isinstance(json, list):
-            return self._new_instance([ self.item_type.from_json(item, models) for item in json ])
-        else:
-            raise DeserializationError("%s expected a list or None, got %s" % (self, json))
-
-    def validate(self, value, detail=True):
-        super(Seq, self).validate(value, True)
-
-        if value is not None:
-            if not (self._is_seq(value) and all(self.item_type.is_valid(item) for item in value)):
-                if self._is_seq(value):
-                    invalid = []
-                    for item in value:
-                        if not self.item_type.is_valid(item):
-                            invalid.append(item)
-                    msg = "" if not detail else "expected an element of %s, got seq with invalid items %r" % (self, invalid)
-                    raise ValueError(msg)
-                else:
-                    msg = "" if not detail else "expected an element of %s, got %r" % (self, value)
-                    raise ValueError(msg)
-
-    @classmethod
-    def _is_seq(cls, value):
-        return ((isinstance(value, collections.Sequence) or cls._is_seq_like(value)) and
-                not isinstance(value, string_types))
-
-    @classmethod
-    def _is_seq_like(cls, value):
-        return (isinstance(value, (collections.Container, collections.Sized, collections.Iterable))
-                and hasattr(value, "__getitem__") # NOTE: this is what makes it disallow set type
-                and not isinstance(value, collections.Mapping))
-
-    def _new_instance(self, value):
-        return value
-
-    def _sphinx_type(self):
-        return self._sphinx_prop_link() + "( %s )" % self.item_type._sphinx_type()
-
-class List(Seq):
-    ''' Accept Python list values.
-
-    '''
-
-    def __init__(self, item_type, default=[], help=None):
-        # todo: refactor to not use mutable objects as default values.
-        # Left in place for now because we want to allow None to express
-        # optional values. Also in Dict.
-        super(List, self).__init__(item_type, default=default, help=help)
-
-    @classmethod
-    def wrap(cls, value):
-        ''' Some property types need to wrap their values in special containers, etc.
-
-        '''
-        if isinstance(value, list):
-            if isinstance(value, PropertyValueList):
-                return value
-            else:
-                return PropertyValueList(value)
-        else:
-            return value
-
-    @classmethod
-    def _is_seq(cls, value):
-        return isinstance(value, list)
-
-class Array(Seq):
-    ''' Accept NumPy array values.
-
-    '''
-
-    @classmethod
-    def _is_seq(cls, value):
-        import numpy as np
-        return isinstance(value, np.ndarray)
-
-    def _new_instance(self, value):
-        import numpy as np
-        return np.array(value)
-
-
-class Dict(ContainerProperty):
-    ''' Accept Python dict values.
-
-    If a default value is passed in, then a shallow copy of it will be
-    used for each new use of this property.
-
-    '''
-
-    def __init__(self, keys_type, values_type, default={}, help=None):
-        self.keys_type = self._validate_type_param(keys_type)
-        self.values_type = self._validate_type_param(values_type)
-        super(Dict, self).__init__(default=default, help=help)
-
-    def __str__(self):
-        return "%s(%s, %s)" % (self.__class__.__name__, self.keys_type, self.values_type)
-
-    @property
-    def type_params(self):
-        return [self.keys_type, self.values_type]
-
-    def from_json(self, json, models=None):
-        if json is None:
-            return None
-        elif isinstance(json, dict):
-            return { self.keys_type.from_json(key, models): self.values_type.from_json(value, models) for key, value in iteritems(json) }
-        else:
-            raise DeserializationError("%s expected a dict or None, got %s" % (self, json))
-
-    def validate(self, value, detail=True):
-        super(Dict, self).validate(value, detail)
-
-        if value is not None:
-            if not (isinstance(value, dict) and \
-                    all(self.keys_type.is_valid(key) and self.values_type.is_valid(val) for key, val in iteritems(value))):
-                msg = "" if not detail else "expected an element of %s, got %r" % (self, value)
-                raise ValueError(msg)
-
-    @classmethod
-    def wrap(cls, value):
-        ''' Some property types need to wrap their values in special containers, etc.
-
-        '''
-        if isinstance(value, dict):
-            if isinstance(value, PropertyValueDict):
-                return value
-            else:
-                return PropertyValueDict(value)
-        else:
-            return value
-
-    def _sphinx_type(self):
-        return self._sphinx_prop_link() + "( %s, %s )" % (self.keys_type._sphinx_type(), self.values_type._sphinx_type())
-
-class ColumnData(Dict):
-    ''' Accept a Python dictionary suitable as the ``data`` attribute of a
-    :class:`~bokeh.models.sources.ColumnDataSource`.
-
-    This class is a specialization of ``Dict`` that handles efficiently
-    encoding columns that are NumPy arrays.
-
-    '''
-
-    def make_descriptors(self, base_name):
-        ''' Return a list of ``ColumnDataPropertyDescriptor`` instances to
-        install on a class, in order to delegate attribute access to this
-        property.
-
-        Args:
-            base_name (str) : the name of the property these descriptors are for
-
-        Returns:
-            list[ColumnDataPropertyDescriptor]
-
-        The descriptors returned are collected by the ``MetaHasProps``
-        metaclass and added to ``HasProps`` subclasses during class creation.
-        '''
-        return [ ColumnDataPropertyDescriptor(base_name, self) ]
-
-
-    def from_json(self, json, models=None):
-        ''' Decodes column source data encoded as lists or base64 strings.
-        '''
-        if json is None:
-            return None
-        elif not isinstance(json, dict):
-            raise DeserializationError("%s expected a dict or None, got %s" % (self, json))
-        new_data = {}
-        for key, value in json.items():
-            key = self.keys_type.from_json(key, models)
-            if isinstance(value, dict) and '__ndarray__' in value:
-                new_data[key] = decode_base64_dict(value)
-            elif isinstance(value, list) and any(isinstance(el, dict) and '__ndarray__' in el for el in value):
-                new_list = []
-                for el in value:
-                    if isinstance(el, dict) and '__ndarray__' in el:
-                        el = decode_base64_dict(el)
-                    elif isinstance(el, list):
-                        el = self.values_type.from_json(el)
-                    new_list.append(el)
-                new_data[key] = new_list
-            else:
-                new_data[key] = self.values_type.from_json(value, models)
-        return new_data
-
-    def serialize_value(self, value):
-        return transform_column_source_data(value)
-
-    @classmethod
-    def wrap(cls, value):
-        ''' Some property types need to wrap their values in special containers, etc.
-
-        '''
-        if isinstance(value, dict):
-            if isinstance(value, PropertyValueColumnData):
-                return value
-            else:
-                return PropertyValueColumnData(value)
-        else:
-            return value
-
-class Tuple(ContainerProperty):
-    ''' Accept Python tuple values.
-
-    '''
-    def __init__(self, tp1, tp2, *type_params, **kwargs):
-        self._type_params = list(map(self._validate_type_param, (tp1, tp2) + type_params))
-        super(Tuple, self).__init__(default=kwargs.get("default"), help=kwargs.get("help"))
-
-    def __str__(self):
-        return "%s(%s)" % (self.__class__.__name__, ", ".join(map(str, self.type_params)))
-
-    @property
-    def type_params(self):
-        return self._type_params
-
-    def from_json(self, json, models=None):
-        if json is None:
-            return None
-        elif isinstance(json, list):
-            return tuple(type_param.from_json(item, models) for type_param, item in zip(self.type_params, json))
-        else:
-            raise DeserializationError("%s expected a list or None, got %s" % (self, json))
-
-    def validate(self, value, detail=True):
-        super(Tuple, self).validate(value, detail)
-
-        if value is not None:
-            if not (isinstance(value, (tuple, list)) and len(self.type_params) == len(value) and \
-                    all(type_param.is_valid(item) for type_param, item in zip(self.type_params, value))):
-                msg = "" if not detail else "expected an element of %s, got %r" % (self, value)
-                raise ValueError(msg)
-
-    def _sphinx_type(self):
-        return self._sphinx_prop_link() + "( %s )" % ", ".join(x._sphinx_type() for x in self.type_params)
-
-class RelativeDelta(Dict):
-    ''' Accept RelativeDelta dicts for time delta values.
-
-    '''
-
-    def __init__(self, default={}, help=None):
-        keys = Enum("years", "months", "days", "hours", "minutes", "seconds", "microseconds")
-        values = Int
-        super(RelativeDelta, self).__init__(keys, values, default=default, help=help)
-
-    def __str__(self):
-        return self.__class__.__name__
+from .property.container import Array; Array
+from .property.container import ColumnData; ColumnData
+from .property.container import Dict; Dict
+from .property.container import List; List
+from .property.container import Seq; Seq
+from .property.container import Tuple; Tuple
+from .property.container import RelativeDelta; RelativeDelta
 
 #------------------------------------------------------------------------------
 # DataSpec properties
@@ -2066,16 +1801,14 @@ _all_props = set(x for x in globals().values() if isinstance(x, type) and issubc
 _all_props.remove(Property)
 _all_props.remove(PrimitiveProperty)
 _all_props.remove(ParameterizedProperty)
-_all_props.remove(ContainerProperty)
 def _find_and_remove(typ):
     global _all_props
     sub = set(x for x in _all_props if issubclass(x, typ))
     _all_props -= sub
     return sub
 _data_specs = "\n".join(sorted(".. autoclass:: %s" % x.__name__ for x in _find_and_remove(DataSpec)))
-_containers = "\n".join(sorted(".. autoclass:: %s" % x.__name__ for x in _find_and_remove(ContainerProperty)))
 _basic = "\n".join(sorted(".. autoclass:: %s" % x.__name__ for x in _all_props))
 
-__doc__ = format_docstring(__doc__, basic_properties=_basic, container_properties=_containers, dataspec_properties=_data_specs)
+__doc__ = format_docstring(__doc__, basic_properties=_basic, dataspec_properties=_data_specs)
 
-del _all_props, _data_specs, _containers, _basic, _find_and_remove
+del _all_props, _data_specs, _basic, _find_and_remove
